+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 ## 1.0.0
+
 - Initial release
+
+## 2.0.0
+
+- Added support for history management
+- Added a new API, less error-prone and more flexible